major refactor of pokeHelpers. Moved its utilities into a separate utils file.
